Revert "BugFix: assign proposition.estCorrect to Long instead of Integer"
This reverts commit 5fc775f7cd291bb6d175bad8028492d00b675580.

+++ src/fr/cdamassy2021/service/QuestionService.java
@@ -81,7 +81,7 @@ public class QuestionService {
     	for(int i = 0; i<allPropositionsLibelles.length; i++)
     	{
     		Proposition prop = new Proposition();
-    		prop.setEstCorrecte(Long.parseLong(allPropositionsCorrectnesses.get(i)));
+    		prop.setEstCorrecte(Integer.parseInt(allPropositionsCorrectnesses.get(i)));
     		prop.setLibelle(allPropositionsLibelles[i]);
     		props.add(prop);
     	}
